# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AstraWatch Unified API",
    "description": "Official OpenAPI 3.0 specification for AstraWatch autonomous telemetry, anomaly detection, incident orchestration, and K8s self-healing microservices.",
    "version": "1.0.0",
    "contact": {
      "name": "AstraWatch Engineering",
      "url": "https://astrawatch.io"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1",
      "description": "Collector Service (:8080)"
    },
    {
      "url": "http://localhost:8000/v1",
      "description": "Analyzer Service (:8000)"
    },
    {
      "url": "http://localhost:8082/api/v1",
      "description": "Orchestrator Service (:8082)"
    },
    {
      "url": "http://localhost:8081/healthz",
      "description": "K8s Operator Service (:8081 healthz)"
    },
    {
      "url": "http://localhost:8084",
      "description": "Realtime Gateway (:8084)"
    },
    {
      "url": "http://localhost:8085/api/v1",
      "description": "Payment Service (:8085)"
    }
  ],
...
  "paths": {
...
    "/api/v1/incidents": {
      "post": {
        "tags": [
          "Orchestrator: Incidents"
        ],
        "summary": "Create Incident",
        "description": "Creates a new incident with severity level, affected service, and description.",
        "operationId": "oi1",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "id": "inc-001",
                  "severity": "HIGH",
                  "state": "OPEN",
                  "createdAt": "2026-01-01T00:00:00Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Incident payload",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "serviceId": "svc-001",
                "anomalyId": "anom-001",
                "severity": "HIGH",
                "title": "Payment service latency spike",
                "description": "P99 latency exceeds 5s threshold"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Orchestrator: Incidents"
        ],
        "summary": "List Incidents",
        "description": "Lists all incidents, optionally filtered by service.",
        "operationId": "oi2",
        "parameters": [
          {
            "name": "serviceId",
            "in": "query",
            "required": false,
            "description": "Filter by service",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": [
                  {
                    "id": "inc-001",
                    "title": "Payment latency",
                    "severity": "HIGH",
                    "state": "OPEN"
                  }
                ]
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/incidents/:id": {
      "get": {
        "tags": [
          "Orchestrator: Incidents"
        ],
        "summary": "Get Incident",
        "description": "Fetches detailed view of a single incident including assignment and state.",
        "operationId": "oi3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Incident ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "id": "inc-001",
                  "serviceId": "svc-001",
                  "severity": "HIGH",
                  "state": "INVESTIGATING",
                  "title": "Payment latency spike",
                  "assignedTo": "usr-001"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/incidents/:id/assign": {
      "post": {
        "tags": [
          "Orchestrator: Incidents"
        ],
        "summary": "Assign Incident",
        "description": "Assigns incident ownership to a user.",
        "operationId": "oi4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Incident ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "id": "inc-001",
                  "assignedTo": "usr-001"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Assignee",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "userId": "usr-001"
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/:id/resolve": {
      "post": {
        "tags": [
          "Orchestrator: Incidents"
        ],
        "summary": "Resolve Incident",
        "description": "Resolves an incident with a resolution note. Transitions state to RESOLVED.",
        "operationId": "oi5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Incident ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "id": "inc-001",
                  "state": "RESOLVED",
                  "resolvedAt": "2026-01-01T01:00:00Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Resolution note",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "resolutionNote": "Rolled back faulty deployment v2.3.1"
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/:id/escalate": {
      "post": {
        "tags": [
          "Orchestrator: Incidents"
        ],
        "summary": "Escalate Incident",
        "description": "Escalates incident to specified user or on-call tier.",
        "operationId": "oi6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Incident ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "id": "inc-001",
                  "state": "ESCALATED"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Escalation details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "escalateTo": "oncall-lead",
                "reason": "No response after 30 minutes"
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/:id/timeline": {
      "get": {
        "tags": [
          "Orchestrator: Incidents"
        ],
        "summary": "Incident Timeline",
        "description": "Fetches the full timeline event log for an incident.",
        "operationId": "oi7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Incident ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": [
                  {
                    "id": 1,
                    "eventType": "CREATED",
                    "payload": "{}",
                    "createdAt": "2026-01-01T00:00:00Z"
                  }
                ]
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/incidents/:id/transition": {
      "patch": {
        "tags": [
          "Orchestrator: Incidents"
        ],
        "summary": "Transition State",
        "description": "Triggers a Spring StateMachine lifecycle event to transition incident state.",
        "operationId": "oi8",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Incident ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "id": "inc-001",
                  "state": "INVESTIGATING"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Lifecycle event",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "event": "INVESTIGATE"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer JWT token issued by POST /api/v1/auth/login"
      }
    },
...
  }
}